Split top_bar code
